requirements.txt
langchain>=0.1.20
langchain-core>=0.1.35
google-generativeai>=0.3.2
pinecone-client>=3.0.0
sentence-transformers>=2.3.1
torch>=2.0.0
pypdf>=3.17.4
